Split files into level-specific and level-generic
